refactor: move and sentralize chart functionality in domain
