+++ src/libecalc/domain/process/chart/chart_area_flag.py
