+++ src/libecalc/domain/process/chart/compressor/__init__.py
