+++ src/libecalc/domain/process/chart/compressor/chart_creator.py
@@ -1,15 +1,15 @@
 import numpy as np
 
 from libecalc.common.serializable_chart import ChartCurveDTO, VariableSpeedChartDTO
-from libecalc.domain.process.compressor.core.train.chart import VariableSpeedCompressorChart
-from libecalc.domain.process.compressor.core.train.chart.generic_chart_data import (
+from libecalc.domain.process.chart.compressor import VariableSpeedCompressorChart
+from libecalc.domain.process.chart.compressor.generic_chart_data import (
     UNIFIED_GENERIC_CHART_CURVE_MAXIMUM_SPEED_HEADS,
     UNIFIED_GENERIC_CHART_CURVE_MAXIMUM_SPEED_RATES,
     UNIFIED_GENERIC_CHART_CURVE_MINIMUM_SPEED_HEADS,
     UNIFIED_GENERIC_CHART_CURVE_MINIMUM_SPEED_RATES,
 )
-from libecalc.domain.process.compressor.core.train.chart.types import CompressorChartResult
-from libecalc.domain.process.compressor.core.train.chart.variable_speed_compressor_chart import (
+from libecalc.domain.process.chart.compressor.types import CompressorChartResult
+from libecalc.domain.process.chart.compressor.variable_speed_compressor_chart import (
     logger,
 )
 from libecalc.domain.process.compressor.core.train.utils.numeric_methods import (
