+++ src/libecalc/domain/process/chart/compressor/compressor_chart_dto.py
@@ -3,7 +3,7 @@ from typing import Annotated, Union
 from pydantic import Field
 
 from libecalc.common.serializable_chart import SingleSpeedChartDTO, VariableSpeedChartDTO
-from libecalc.domain.process.dto.chart import (
+from libecalc.domain.process.chart.generic.chart import (
     GenericChartFromDesignPoint,
     GenericChartFromInput,
 )
