+++ src/libecalc/domain/process/chart/compressor/generic_chart_data.py
