+++ src/libecalc/domain/process/chart/compressor/single_speed_compressor_chart.py
@@ -1,10 +1,10 @@
 from copy import deepcopy
 
-from libecalc.domain.process.compressor.core.train.chart.types import (
+from libecalc.domain.process.chart import SingleSpeedChart
+from libecalc.domain.process.chart.chart_area_flag import ChartAreaFlag
+from libecalc.domain.process.chart.compressor.types import (
     CompressorChartHeadEfficiencyResultSinglePoint,
 )
-from libecalc.domain.process.core.chart import SingleSpeedChart
-from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
 
 
 class SingleSpeedCompressorChart(SingleSpeedChart):
