+++ src/libecalc/domain/process/chart/compressor/types.py
@@ -1,4 +1,4 @@
-from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
+from libecalc.domain.process.chart.chart_area_flag import ChartAreaFlag
 
 
 class CompressorChartHeadEfficiencyResultSinglePoint:
