+++ src/libecalc/domain/process/chart/generic/__init__.py
@@ -1,0 +1,4 @@
+from .chart import (
+    GenericChartFromDesignPoint,
+    GenericChartFromInput,
+)
