+++ src/libecalc/domain/process/compressor/core/results.py
@@ -5,7 +5,7 @@ import numpy as np
 from libecalc.common.fluid import FluidComposition, FluidStream
 from libecalc.common.serializable_chart import SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.common.units import Unit
-from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
+from libecalc.domain.process.chart.chart_area_flag import ChartAreaFlag
 from libecalc.domain.process.core.results.compressor import (
     CompressorStageResult,
     CompressorStreamCondition,
