+++ src/libecalc/domain/process/compressor/core/sampled/compressor_model_sampled.py
@@ -11,6 +11,7 @@ from libecalc.common.list.adjustment import transform_linear
 from libecalc.common.list.list_utils import array_to_list
 from libecalc.common.logger import logger
 from libecalc.common.units import Unit
+from libecalc.domain.process.chart.chart_area_flag import ChartAreaFlag
 from libecalc.domain.process.compressor.core.base import CompressorModel
 from libecalc.domain.process.compressor.core.sampled.compressor_model_sampled_1d import (
     CompressorModelSampled1D,
@@ -31,7 +32,6 @@ from libecalc.domain.process.compressor.core.sampled.constants import (
     RATE_NAME,
 )
 from libecalc.domain.process.compressor.dto import CompressorSampled
-from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
 from libecalc.domain.process.core.results import (
     CompressorStageResult,
     CompressorStreamCondition,
