+++ src/libecalc/domain/process/compressor/core/train/utils/variable_speed_compressor_train_common_shaft.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from libecalc.common.serializable_chart import ChartCurveDTO
-from libecalc.domain.process.compressor.core.train.chart import (
+from libecalc.domain.process.chart.compressor import (
     SingleSpeedCompressorChart,
     VariableSpeedCompressorChart,
 )
