+++ src/libecalc/domain/process/compressor/core/utils.py
@@ -1,21 +1,22 @@
 from libecalc.common.serializable_chart import SingleSpeedChartDTO, VariableSpeedChartDTO
-from libecalc.domain.process.compressor.core.train.chart import (
+from libecalc.domain.process.chart.compressor import (
     SingleSpeedCompressorChart,
     VariableSpeedCompressorChart,
 )
-from libecalc.domain.process.compressor.core.train.chart.chart_creator import (
+from libecalc.domain.process.chart.compressor.chart_creator import (
     CompressorChartCreator,
 )
+from libecalc.domain.process.chart.compressor.compressor_chart_dto import CompressorChart
+from libecalc.domain.process.chart.generic import (
+    GenericChartFromDesignPoint,
+    GenericChartFromInput,
+)
+from libecalc.domain.process.chart.generic import GenericChartFromInput as GenericChartFromInputDTO
 from libecalc.domain.process.compressor.core.train.stage import (
     CompressorTrainStage,
     UndefinedCompressorStage,
 )
-from libecalc.domain.process.compressor.dto import CompressorChart, CompressorStage
-from libecalc.domain.process.dto import (
-    GenericChartFromDesignPoint,
-    GenericChartFromInput,
-)
-from libecalc.domain.process.dto import GenericChartFromInput as GenericChartFromInputDTO
+from libecalc.domain.process.compressor.dto import CompressorStage
 
 
 def _create_compressor_chart(
