+++ src/libecalc/domain/process/compressor/dto/__init__.py
@@ -1,4 +1,3 @@
-from .chart import CompressorChart
 from .compressor_consumer_function import CompressorConsumerFunction
 from .sampled import CompressorSampled
 from .stage import CompressorStage, InterstagePressureControl, MultipleStreamsCompressorStage
