+++ src/libecalc/domain/process/dto/__init__.py
@@ -1,8 +1,4 @@
 from .base import ConsumerFunction, EnergyModel
-from .chart import (
-    GenericChartFromDesignPoint,
-    GenericChartFromInput,
-)
 from .direct import DirectConsumerFunction
 from .sampled import EnergyModelSampled
 from .tabulated import TabulatedConsumerFunction, TabulatedData, Variables
