+++ src/libecalc/domain/process/pump/factory.py
@@ -5,7 +5,7 @@ from typing import Any
 from libecalc.common.chart_type import ChartType
 from libecalc.common.logger import logger
 from libecalc.common.serializable_chart import ChartCurveDTO, SingleSpeedChartDTO, VariableSpeedChartDTO
-from libecalc.domain.process.core.chart import SingleSpeedChart, VariableSpeedChart
+from libecalc.domain.process.chart import SingleSpeedChart, VariableSpeedChart
 from libecalc.domain.process.pump.pump import PumpModel, PumpModelDTO, PumpSingleSpeed, PumpVariableSpeed
 
 # def evaluate_streams(
