+++ src/libecalc/presentation/yaml/mappers/model.py
@@ -6,8 +6,9 @@ from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureContr
 from libecalc.common.fluid import MultipleStreamsAndPressureStream
 from libecalc.common.serializable_chart import ChartCurveDTO, SingleSpeedChartDTO, VariableSpeedChartDTO
 from libecalc.common.units import Unit
+from libecalc.domain.process.chart.compressor.compressor_chart_dto import CompressorChart
+from libecalc.domain.process.chart.generic import GenericChartFromDesignPoint, GenericChartFromInput
 from libecalc.domain.process.compressor.dto import (
-    CompressorChart,
     CompressorStage,
     CompressorTrainSimplifiedWithKnownStages,
     CompressorTrainSimplifiedWithUnknownStages,
@@ -20,8 +21,6 @@ from libecalc.domain.process.compressor.dto import (
 )
 from libecalc.domain.process.dto import (
     EnergyModel,
-    GenericChartFromDesignPoint,
-    GenericChartFromInput,
     Turbine,
 )
 from libecalc.domain.resource import Resource, Resources
