+++ tests/libecalc/core/conftest.py
@@ -10,7 +10,7 @@ import libecalc.dto.fuel_type
 from libecalc.dto.emission import Emission
 from libecalc.domain.process.compressor import dto
 from libecalc.common.fluid import FluidComposition, FluidModel
-from libecalc.domain.process.core.chart import SingleSpeedChart, VariableSpeedChart
+from libecalc.domain.process.chart import SingleSpeedChart, VariableSpeedChart
 from libecalc.domain.process.compressor.core.sampled import CompressorModelSampled
 from libecalc.domain.process.pump.pump import PumpSingleSpeed, PumpVariableSpeed
 from libecalc.domain.process.core.turbine import TurbineModel
