+++ tests/libecalc/core/consumers/system/test_consumer_system.py
@@ -24,7 +24,7 @@ from libecalc.domain.infrastructure.energy_components.legacy_consumer.system.ope
     PumpSystemOperationalSettingExpressions,
 )
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.system.types import ConsumerSystemComponent
-from libecalc.domain.process.core.chart import SingleSpeedChart
+from libecalc.domain.process.chart import SingleSpeedChart
 from libecalc.domain.process.pump.pump import PumpSingleSpeed
 from libecalc.expression import Expression
