+++ tests/libecalc/core/models/chart/test_chart_curve.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from libecalc.common.serializable_chart import ChartCurveDTO
-from libecalc.domain.process.core.chart import ChartCurve
+from libecalc.domain.process.chart import ChartCurve
 
 
 @pytest.fixture
