+++ tests/libecalc/core/models/compressor_modelling/conftest.py
@@ -12,7 +12,7 @@ from libecalc.domain.process.compressor import dto
 from libecalc.common.fixed_speed_pressure_control import FixedSpeedPressureControl
 from libecalc.common.fluid import FluidComposition, FluidModel, MultipleStreamsAndPressureStream
 from libecalc.common.units import Unit
-from libecalc.domain.process.compressor.core.train.chart import SingleSpeedCompressorChart
+from libecalc.domain.process.chart.compressor import SingleSpeedCompressorChart
 from libecalc.domain.process.compressor.core.train.fluid import FluidStream
 from libecalc.domain.process.compressor.core.train.single_speed_compressor_train_common_shaft import (
     SingleSpeedCompressorTrainCommonShaft,
