+++ tests/libecalc/core/models/compressor_modelling/test_compressor_chart.py
@@ -4,8 +4,8 @@ from pytest import approx
 
 from libecalc.common.errors.exceptions import IllegalStateException
 from libecalc.common.serializable_chart import ChartCurveDTO, VariableSpeedChartDTO
-from libecalc.domain.process.core.chart.chart_area_flag import ChartAreaFlag
-from libecalc.domain.process.compressor.core.train.chart import (
+from libecalc.domain.process.chart.chart_area_flag import ChartAreaFlag
+from libecalc.domain.process.chart.compressor import (
     SingleSpeedCompressorChart,
     VariableSpeedCompressorChart,
 )
