+++ tests/libecalc/core/models/compressor_modelling/test_compressor_model_vs_unisim.py
@@ -3,7 +3,7 @@ import pytest
 from pytest import approx
 
 from libecalc.domain.process.compressor import dto
-from libecalc.domain.process.dto import GenericChartFromDesignPoint
+from libecalc.domain.process.chart.generic import GenericChartFromDesignPoint
 from libecalc.common.fluid import EoSModel, FluidComposition, FluidModel
 from libecalc.common.units import Unit
 from libecalc.domain.process.compressor.core.train.fluid import FluidStream
